fix: fetch results cached st not getting latest data
