+++ src/features/applications/hooks/useApplicationByTxHash.ts
@@ -1,7 +1,7 @@
 import { api } from "~/utils/api";
 
 import type { UseTRPCQueryResult } from "@trpc/react-query/shared";
-import type { Attestation } from "~/utils/fetchAttestations";
+import type { Attestation } from "~/utils/types";
 
 export function useApplicationByTxHash(transactionId: string): UseTRPCQueryResult<Attestation, unknown> {
   return api.projects.getByTransactionId.useQuery({ transactionId });
