+++ src/features/applications/hooks/useApprovedApplications.ts
@@ -1,7 +1,7 @@
 import { api } from "~/utils/api";
 
 import type { UseTRPCQueryResult } from "@trpc/react-query/shared";
-import type { Attestation } from "~/utils/fetchAttestations";
+import type { Attestation } from "~/utils/types";
 
 export function useApprovedApplications(ids?: string[]): UseTRPCQueryResult<Attestation[], unknown> {
   return api.applications.approvals.useQuery({ ids });
